Renamed parsing to fuckmake and moved the files to core
